+++ src/main/java/org/example/springmvccrud/service/UserService.java
@@ -12,9 +12,9 @@ public interface UserService {
 
     void updateUser(User user);
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 
     List<User> getAllUsers();
 }
